Added a new template under renderers.
The new template "SetupListHTML.active.template.html" is an attempt to
optimize the solution used in SetupListHTML.template.html.

Is is marked as "active" bvecause it does not rely on the "VenuePlan"
specification, instead all data is retrieved using calls specified in
the template.

The upside of this is that it easier to optimize, since all the
information used to generate the list is specified in the template
itself. This makes it easier to maintain and to optimize.