Implemented caching of CRM content.
Turns out that AMSoftware.CRM uses a lazy loading model for data on
Linked Entities, which is what caused the obscenely long rendering times
previously.

Added lines to make sure that all  data is present in local memory and
strip out the lazy loading logic from the results of GetContent.